LayoutTests/svg/animations/svgenum-animation-12-expected.txt

SVG 1.1 dynamic animation tests

Test SVGSpreadMethodType enumeration animations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS gradient.spreadMethod.animVal is SVGGradientElement.SVG_SPREADMETHOD_PAD
PASS gradient.spreadMethod.baseVal is SVGGradientElement.SVG_SPREADMETHOD_PAD
PASS gradient.spreadMethod.animVal is SVGGradientElement.SVG_SPREADMETHOD_PAD
PASS gradient.spreadMethod.baseVal is SVGGradientElement.SVG_SPREADMETHOD_PAD
PASS gradient.spreadMethod.animVal is SVGGradientElement.SVG_SPREADMETHOD_PAD
PASS gradient.spreadMethod.baseVal is SVGGradientElement.SVG_SPREADMETHOD_PAD
PASS gradient.spreadMethod.animVal is SVGGradientElement.SVG_SPREADMETHOD_REFLECT
PASS gradient.spreadMethod.baseVal is SVGGradientElement.SVG_SPREADMETHOD_PAD
PASS gradient.spreadMethod.animVal is SVGGradientElement.SVG_SPREADMETHOD_REFLECT
PASS gradient.spreadMethod.baseVal is SVGGradientElement.SVG_SPREADMETHOD_PAD
PASS gradient.spreadMethod.animVal is SVGGradientElement.SVG_SPREADMETHOD_REPEAT
PASS gradient.spreadMethod.baseVal is SVGGradientElement.SVG_SPREADMETHOD_PAD
PASS gradient.spreadMethod.animVal is SVGGradientElement.SVG_SPREADMETHOD_REPEAT
PASS gradient.spreadMethod.baseVal is SVGGradientElement.SVG_SPREADMETHOD_PAD
PASS gradient.spreadMethod.animVal is SVGGradientElement.SVG_SPREADMETHOD_REPEAT
PASS gradient.spreadMethod.baseVal is SVGGradientElement.SVG_SPREADMETHOD_PAD
PASS successfullyParsed is true

TEST COMPLETE
